Use classes for audio dependencies installations
